started on parsing; added a lot of t_type memebers
some of the t_type toke types need to be adressed in lexer. others are
constructed from the base types that lexer finds. they are contructed or
get asigned more specific roles in parsing part. info in IMPORTANT
NOTES.
parser: did <word> and <string> but needs to be checked because some
cases might be missing