added pwd, join_wrapper, and prompt_check
now the prompt show like this: minishel $USER absolute path
so it look more like a normal bash prompt

i added a wrapper for ft_strjoin. so now its possible to pass allocated
strings and free them. its in safe_ft.c

did the pwd function - its very simple

changed a bit Makefile

parser still needs to be checked but i am a bit lost as to how to
implement it in the best way