make works. added the executor part;not tested
there are some possible complications in how the executor is handled. as
it should be able to manage multiple in/outfiles whic for now it does
nto. but i am thinking of first making a working basic version of
executor so that later it could be expanded further. baby steps!