corrected parser. should be almost perfect now i think
made so that when things are in quotes then the nodes with qoutes get
deleted. so the nodes are clean and the WHITESPACE divides each necesary
part

also when expander is working. if you expand something that does not
exist it deletes the node, so it doesnt get proccessed.

still need to do the part after the parser and before executor where we
send the expanded nodes through parser again so we can decide what is
their role